+++ plugins/device_preview_screenshot/lib/device_preview_screenshot.dart
@@ -1,0 +1,9 @@
+library device_preview_screenshot;
+
+export 'package:device_preview/device_preview.dart';
+
+export 'src/processors/base64.dart';
+export 'src/processors/file.dart';
+export 'src/processors/processor.dart';
+
+export 'src/section.dart';